Move PyRef to pyobject module
